% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.r
\name{as_chunked_array}
\alias{as_chunked_array}
\title{Convert to chunked_array}
\usage{
as_chunked_array(x)
}
\arguments{
\item{x}{a vector or chunked_array object}
}
\value{
A chunked_array object
}
\description{
Convert a numeric or character vector to a chunked_array object.
If the input is already a chunked_array, it is returned as-is.
}
