% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.r
\name{c2}
\alias{c2}
\title{Concatenate Chunked Arrays}
\usage{
c2(x, y)
}
\arguments{
\item{x}{a vector or chunked_array object}

\item{y}{a vector or chunked_array object}
}
\value{
A chunked_array object
}
\description{
Concatenate vectors into a chunked array
}
\details{
Combine two vector or chunked_array objects into a single
chunked_array without copying the underlying data.
}
\examples{
a <- c2(1:100, 101:200)
length(a)
a[150]
}
\author{
Guangchuang Yu
}
