% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{check_pkg}
\alias{check_pkg}
\title{check_pkg}
\usage{
check_pkg(pkg, reason = NULL, ...)
}
\arguments{
\item{pkg}{package names}

\item{reason}{the reason to check the pkg. If NULL, it will set the reason to the parent call.}

\item{...}{additional parameters that passed to \code{rlang::check_installed()}}
}
\value{
see also \link[rlang:is_installed]{check_installed}
}
\description{
Check whether the input packages are installed
}
\details{
This function check whether the input packages are installed. If not, it asks the user whether to install the missing packages.
}
\author{
Guangchuang Yu
}
