% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations}
\alias{combinations}
\title{combinations}
\usage{
combinations(n)
}
\arguments{
\item{n}{number of sets}
}
\value{
a list of all combinations
}
\description{
all possible combinations of n sets
}
