% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_yulab_file}
\alias{download_yulab_file}
\title{Process YuLab File Download}
\usage{
download_yulab_file(destfile, urls, gzfile = FALSE, appname = NULL)
}
\arguments{
\item{destfile}{character. Local file path.}

\item{urls}{character vector. Base URLs for download. Remote is is \code{urls/basename(destfile)}}

\item{gzfile}{logical. Whether the remote file is gzipped.}

\item{appname}{character. R package name.}
}
\description{
Process YuLab File Download
}
\author{
Guangchuang Yu
}
