% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{get_dependencies}
\alias{get_dependencies}
\title{get_dependencies}
\usage{
get_dependencies(pkg, repo = c("CRAN", "BioC"))
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
Get reverse dependencies
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=Githubpkg]{Githubpkg()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=is.installed]{is.installed()}},
\code{\link[=mypkg]{mypkg()}},
\code{\link[=packageTitle]{packageTitle()}},
\code{\link[=pload]{pload()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
