% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{Githubpkg}
\alias{Githubpkg}
\title{Markdown link to GitHub}
\usage{
Githubpkg(user, pkg)
}
\arguments{
\item{user}{github user}

\item{pkg}{package name}
}
\value{
md text string
}
\description{
Markdown link to GitHub
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=get_dependencies]{get_dependencies()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=is.installed]{is.installed()}},
\code{\link[=mypkg]{mypkg()}},
\code{\link[=packageTitle]{packageTitle()}},
\code{\link[=pload]{pload()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
