% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{Githubpkg}
\alias{Githubpkg}
\title{Githubpkg}
\usage{
Githubpkg(user, pkg)
}
\arguments{
\item{user}{github user}

\item{pkg}{package name}
}
\value{
string
}
\description{
print md text of package with link to github repo
}
\author{
Guangchuang Yu
}
