% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{has_internet}
\alias{has_internet}
\title{has_internet}
\usage{
has_internet(site = "https://www.baidu.com/")
}
\arguments{
\item{site}{URL to test connection}
}
\value{
logical value
}
\description{
test for internect connection via reading lines from a URL
}
\seealso{
Other os-utils:
\code{\link[=exec]{exec()}},
\code{\link[=user_dir]{user_dir()}}
}
\author{
Guangchuang Yu
}
\concept{os-utils}
