% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_zip.R
\name{install_zip}
\alias{install_zip}
\title{install_zip}
\usage{
install_zip(file, args = "--no-build-vignettes")
}
\arguments{
\item{file}{zip file}

\item{args}{argument to build package}
}
\value{
No return value, called for install R package from zip file of source codes
}
\description{
install R package from zip file of source codes
}
\author{
Guangchuang Yu
}
