% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_zip.R
\name{install_zip_gh}
\alias{install_zip_gh}
\title{install_zip_gh}
\usage{
install_zip_gh(repo, ref = "master", args = "--no-build-vignettes")
}
\arguments{
\item{repo}{github repo}

\item{ref}{github branch, default is master}

\item{args}{argument to build package}
}
\value{
No return value, called for installing github package
}
\description{
install github package
}
\details{
it download the zip file first and use \code{install_zip} to install it
}
\author{
Guangchuang Yu
}
