% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{is.installed}
\alias{is.installed}
\title{is.installed}
\usage{
is.installed(packages)
}
\arguments{
\item{packages}{package names}
}
\value{
logical vector
}
\description{
Check whether packages are installed
}
\examples{
is.installed(c("dplyr", "ggplot2"))
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=Githubpkg]{Githubpkg()}},
\code{\link[=get_dependencies]{get_dependencies()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=mypkg]{mypkg()}},
\code{\link[=packageTitle]{packageTitle()}},
\code{\link[=pload]{pload()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
