% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls2df}
\alias{ls2df}
\title{Convert a list of vectors (e.g., gene IDs) to \code{data.frame}}
\usage{
ls2df(inputList)
}
\arguments{
\item{inputList}{List of vectors}
}
\value{
\code{data.frame}
}
\description{
Convert a list of vectors to a data.frame
}
