% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-utils.R
\name{mat2list}
\alias{mat2list}
\title{mat2list}
\usage{
mat2list(x)
}
\arguments{
\item{x}{the input matrix}
}
\value{
a list that contains matrix columns as its elements
}
\description{
convert a matrix to a list
}
\examples{
x <- matrix(1:15, nrow = 3)
mat2list(x)
}
