% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{mypkg}
\alias{mypkg}
\title{mypkg}
\usage{
mypkg(pkg, url)
}
\arguments{
\item{pkg}{package name}

\item{url}{package url}
}
\value{
md text string
}
\description{
Markdown link to a package
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=Githubpkg]{Githubpkg()}},
\code{\link[=get_dependencies]{get_dependencies()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=is.installed]{is.installed()}},
\code{\link[=packageTitle]{packageTitle()}},
\code{\link[=pload]{pload()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
