% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{packageTitle}
\alias{packageTitle}
\title{packageTitle}
\usage{
packageTitle(pkg, repo = "CRAN")
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
Extract package title
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=Githubpkg]{Githubpkg()}},
\code{\link[=get_dependencies]{get_dependencies()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=is.installed]{is.installed()}},
\code{\link[=mypkg]{mypkg()}},
\code{\link[=pload]{pload()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
