% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ratio.R
\name{parse_ratio}
\alias{parse_ratio}
\title{parse_ratio}
\usage{
parse_ratio(ratio)
}
\arguments{
\item{ratio}{Character vector of ratios}
}
\value{
Numeric vector
}
\description{
Parse character ratio to double, e.g., \code{1/5} → \code{0.2}
}
\author{
Guangchuang Yu
}
