% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{pload}
\alias{pload}
\title{pload}
\usage{
pload(package, action = "auto")
}
\arguments{
\item{package}{package name}

\item{action}{Installation function; \code{"auto"} tries \code{BiocManager::install()} if available}
}
\value{
the selected package loaded to the R session
}
\description{
Load a package
}
\details{
Uses \code{library()} to load \code{package}. If not installed, attempts installation
via \code{rlang::check_installed()} (optionally using \code{BiocManager::install()}).
}
\seealso{
Other pkg-utils:
\code{\link[=CRANpkg]{CRANpkg()}},
\code{\link[=Githubpkg]{Githubpkg()}},
\code{\link[=get_dependencies]{get_dependencies()}},
\code{\link[=get_fun_from_pkg]{get_fun_from_pkg()}},
\code{\link[=is.installed]{is.installed()}},
\code{\link[=mypkg]{mypkg()}},
\code{\link[=packageTitle]{packageTitle()}}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
