% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{rbindlist}
\alias{rbindlist}
\title{rbindlist}
\usage{
rbindlist(x)
}
\arguments{
\item{x}{List with similar elements that can be row-bound}
}
\value{
\code{data.frame}
}
\description{
Row-bind a list
}
\author{
Guangchuang Yu
}
