% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_tsv_with_cache}
\alias{read_tsv_with_cache}
\alias{read_with_cache}
\title{read_with_cache}
\usage{
read_tsv_with_cache(
  file_url,
  reader = utils::read.delim,
  params = list(sep = "\\t", header = FALSE),
  cache_dir = tempdir()
)

read_with_cache(
  file_url,
  reader = readLines,
  params = list(),
  cache_dir = tempdir()
)
}
\arguments{
\item{file_url}{a file or url}

\item{reader}{a function to read the 'file_url'}

\item{params}{a list of parameters that passed to the 'reader'}

\item{cache_dir}{a folder to store cache files}
}
\value{
the output of using the 'reader' to read the 'file_url' with parameters specified by the 'params'
}
\description{
read file with caching
}
\details{
This function read a file (local or url) and cache the content.
}
\author{
Yonghe Xia and Guangchuang Yu
}
