% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_range}
\alias{scale_range}
\title{scale-range}
\usage{
scale_range(data)
}
\arguments{
\item{data}{the input data.}
}
\value{
normalized data
}
\description{
normalized data by range
}
\author{
Guangchuang Yu
}
