% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scihub-dl.R
\name{scihub_dl}
\alias{scihub_dl}
\title{download publication via scihub}
\usage{
scihub_dl(doi, scihub = "sci-hub.tw", download = TRUE)
}
\arguments{
\item{doi}{doi}

\item{scihub}{scihub website}

\item{download}{whether download the pdf file}
}
\value{
pdf url
}
\description{
using scihub to download publication using doi
}
\author{
Guangchuang Yu
}
