% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_detect}
\alias{str_detect}
\title{str_detect}
\usage{
str_detect(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{Input string}

\item{pattern}{Pattern to look for}

\item{negate}{If \code{TRUE}, invert the result}
}
\value{
Logical vector
}
\description{
Detect presence/absence of a match
}
\seealso{
Other str-utils:
\code{\link[=str_extract]{str_extract()}},
\code{\link[=str_starts]{str_starts()}},
\code{\link[=str_wrap]{str_wrap()}}
}
\author{
Guangchuang Yu
}
\concept{str-utils}
