% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_wrap}
\alias{str_wrap}
\title{str_wrap}
\usage{
str_wrap(string, width = getOption("width"))
}
\arguments{
\item{string}{Input string}

\item{width}{Maximum characters before wrapping}
}
\value{
Updated strings with \code{"\\n"} inserted
}
\description{
Wrap long strings to multiple lines
}
\seealso{
Other str-utils:
\code{\link[=str_detect]{str_detect()}},
\code{\link[=str_extract]{str_extract()}},
\code{\link[=str_starts]{str_starts()}}
}
\author{
Guangchuang Yu
}
\concept{str-utils}
