% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yulab-utils-package.R
\docType{package}
\name{yulab.utils-package}
\alias{yulab.utils}
\alias{yulab.utils-package}
\title{yulab.utils: Supporting Functions for Packages Maintained by 'YuLab-SMU'}
\description{
Miscellaneous functions commonly used by 'YuLab-SMU'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/yulab.utils/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})
}

}
\keyword{internal}
