% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-utils.r
\name{yulab_inform}
\alias{yulab_inform}
\title{Standardized info function}
\usage{
yulab_inform(message, class = "yulab_info", ...)
}
\arguments{
\item{message}{Information message to display}

\item{class}{Custom info class for categorization}

\item{...}{Additional context information}
}
\value{
No return value, displays information
}
\description{
Standardized info function
}
