% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yulab-msg.R
\name{yulab_msg}
\alias{yulab_msg}
\title{yulab_msg}
\usage{
yulab_msg(pkgname = NULL, n = 1)
}
\arguments{
\item{pkgname}{package name}

\item{n}{number of citation messages}
}
\value{
package message
}
\description{
Messages for R package developed by YuLab
}
\author{
Guangchuang Yu
}
