% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-utils.r
\name{yulab_warn}
\alias{yulab_warn}
\title{Standardized warning function}
\usage{
yulab_warn(message, class = "yulab_warning", ...)
}
\arguments{
\item{message}{Warning message to display}

\item{class}{Custom warning class for categorization}

\item{...}{Additional context information}
}
\value{
No return value, issues warning
}
\description{
Standardized warning function
}
