% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.r
\name{c2}
\alias{c2}
\title{c2}
\usage{
c2(x, y)
}
\arguments{
\item{x}{a vector or chunked_array object}

\item{y}{a vector or chunked_array object}
}
\value{
chunked_array object
}
\description{
chunked array
}
\details{
concate two vector/chunked_array into a chunked_array object
}
\author{
Guangchuang Yu
}
