% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{CRANpkg}
\alias{CRANpkg}
\alias{Biocpkg}
\title{print md text of package with link to homepage (CRAN or Bioconductor)}
\usage{
CRANpkg(pkg)

Biocpkg(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
md text string
}
\description{
print md text of package with link to homepage (CRAN or Bioconductor)
}
\author{
Guangchuang Yu
}
