% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{exec}
\alias{exec}
\title{exec}
\usage{
exec(command)
}
\arguments{
\item{command}{system command to run}
}
\value{
An \code{exec} instance that stores system command outputs
}
\description{
run system command
}
\seealso{
Other os-utils: 
\code{\link{has_internet}()},
\code{\link{user_dir}()}
}
\author{
Guangchuang Yu
}
\concept{os-utils}
