% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{get_dependencies}
\alias{get_dependencies}
\title{get_dependencies}
\usage{
get_dependencies(pkg, repo = c("CRAN", "BioC"))
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
get reverse dependencies
}
\author{
Guangchuang Yu
}
