% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{get_dependencies}
\alias{get_dependencies}
\title{get_dependencies}
\usage{
get_dependencies(pkg, repo = c("CRAN", "BioC"))
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
Get reverse dependencies
}
\seealso{
Other pkg-utils: 
\code{\link{CRANpkg}()},
\code{\link{Githubpkg}()},
\code{\link{get_fun_from_pkg}()},
\code{\link{is.installed}()},
\code{\link{mypkg}()},
\code{\link{packageTitle}()},
\code{\link{pload}()}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
