% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{get_fun_from_pkg}
\alias{get_fun_from_pkg}
\title{get_fun_from_pkg}
\usage{
get_fun_from_pkg(pkg, fun)
}
\arguments{
\item{pkg}{package}

\item{fun}{function}
}
\value{
function
}
\description{
load function from package
}
\examples{
get_fun_from_pkg('utils', 'zip')
}
\seealso{
Other pkg-utils: 
\code{\link{CRANpkg}()},
\code{\link{Githubpkg}()},
\code{\link{get_dependencies}()},
\code{\link{is.installed}()},
\code{\link{mypkg}()},
\code{\link{packageTitle}()}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
