% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{get_fun_from_pkg}
\alias{get_fun_from_pkg}
\title{get_fun_from_pkg}
\usage{
get_fun_from_pkg(pkg, fun)
}
\arguments{
\item{pkg}{package}

\item{fun}{function}
}
\value{
function
}
\description{
load function from package
}
\examples{
get_fun_from_pkg('utils', 'zip')
}
\author{
Guangchuang Yu
}
