% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{is.installed}
\alias{is.installed}
\title{is.installed}
\usage{
is.installed(packages)
}
\arguments{
\item{packages}{package names}
}
\value{
logical vector
}
\description{
Check whether the input packages are installed
}
\details{
This function check whether the input packages are installed
}
\examples{
is.installed(c("dplyr", "ggplot2"))
}
\author{
Guangchuang Yu
}
