% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ls2df}
\alias{ls2df}
\title{Convert a list of vector (e.g, gene IDs) to a data.frame object}
\usage{
ls2df(inputList)
}
\arguments{
\item{inputList}{A list of vector}
}
\value{
a data.frame object.
}
\description{
Convert a list of vector to a data.frame object.
}
