% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{mypkg}
\alias{mypkg}
\title{mypkg}
\usage{
mypkg(pkg, url)
}
\arguments{
\item{pkg}{package name}

\item{url}{package url}
}
\value{
md text string
}
\description{
print md text of link to a pakcage
}
\author{
Guangchuang Yu
}
