% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{o}
\alias{o}
\title{o}
\usage{
o(file = ".")
}
\arguments{
\item{file}{to be open; open workding directory by default}
}
\description{
open working directory
}
\examples{
\dontrun{
## to open current working directory
o()
}
}
\author{
Guangchuang Yu
}
