% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{packageTitle}
\alias{packageTitle}
\title{packageTitle}
\usage{
packageTitle(pkg, repo = "CRAN")
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
Extract package title
}
\author{
Guangchuang Yu
}
