% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{packageTitle}
\alias{packageTitle}
\title{packageTitle}
\usage{
packageTitle(pkg, repo = "CRAN")
}
\arguments{
\item{pkg}{package name}

\item{repo}{'CRAN' and/or 'BioC'}
}
\value{
reverse dependencies
}
\description{
Extract package title
}
\seealso{
Other pkg-utils: 
\code{\link{CRANpkg}()},
\code{\link{Githubpkg}()},
\code{\link{get_dependencies}()},
\code{\link{get_fun_from_pkg}()},
\code{\link{is.installed}()},
\code{\link{mypkg}()},
\code{\link{pload}()}
}
\author{
Guangchuang Yu
}
\concept{pkg-utils}
