% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{pload}
\alias{pload}
\title{pload}
\usage{
pload(package, action = "auto")
}
\arguments{
\item{package}{package name}

\item{action}{Installation function; \code{"auto"} tries \code{BiocManager::install()} if available}
}
\value{
the selected package loaded to the R session
}
\description{
Load a package
}
\author{
Guangchuang Yu
}
\concept{pkg-utils

Uses \code{library()} to load \code{package}. If not installed, attempts installation
via \code{rlang::check_installed()} (optionally using \code{BiocManager::install()}).}
