% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{pload}
\alias{pload}
\title{pload}
\usage{
pload(package, action = "auto")
}
\arguments{
\item{package}{package name}

\item{action}{function used to install package.
If 'action = "auto"', it will try to use 'BiocManager::install()' if it is available.}
}
\value{
the selected package loaded to the R session
}
\description{
loading a package
}
\details{
The function use 'library()' to load the package.
If the package is not installed, the function will try to install it before loading it.
}
\author{
Guangchuang Yu
}
