% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{rbindlist}
\alias{rbindlist}
\title{rbindlist}
\usage{
rbindlist(x)
}
\arguments{
\item{x}{a list that have similar elements that can be rbind to a data.frame}
}
\value{
data.frame
}
\description{
rbind a list
}
\author{
Guangchuang Yu
}
