% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read.cb}
\alias{read.cb}
\title{read.cb}
\usage{
read.cb(reader = read.table, ...)
}
\arguments{
\item{reader}{function to read the clipboard}

\item{...}{parameters for the reader}
}
\value{
clipboard content, output type depends on the output of the reader
}
\description{
read clipboard
}
\seealso{
Other io-utils: 
\code{\link{o}()},
\code{\link{show_in_excel}()},
\code{\link{yread_tsv}()}
}
\author{
Guangchuang Yu
}
\concept{io-utils}
