% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{show_in_excel}
\alias{show_in_excel}
\title{show_in_excel}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{a data frame to be open}
}
\value{
original .data
}
\description{
Open data frame in Excel. It can be used in pipe.
}
\author{
Guangchuang Yu
}
