% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_detect}
\alias{str_detect}
\title{str_detect}
\usage{
str_detect(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{input string}

\item{pattern}{pattern to look for}

\item{negate}{if TRUE, inverts the resulting boolen vector}
}
\value{
logical vector
}
\description{
Detect the presentce/absence of a match
}
\author{
Guangchuang Yu
}
