% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_extract}
\alias{str_extract}
\title{str_extract}
\usage{
str_extract(string, pattern)
}
\arguments{
\item{string}{input string}

\item{pattern}{a regular expression to describe the pattern to extracted from the 'string'}
}
\value{
substring
}
\description{
Extract a substring using a pattern
}
\author{
Guangchuang Yu
}
