% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_extract}
\alias{str_extract}
\title{str_extract}
\usage{
str_extract(string, pattern)
}
\arguments{
\item{string}{Input string}

\item{pattern}{Regular expression to extract}
}
\value{
Substring
}
\description{
Extract a substring using a pattern
}
\seealso{
Other str-utils: 
\code{\link{str_detect}()},
\code{\link{str_starts}()},
\code{\link{str_wrap}()}
}
\author{
Guangchuang Yu
}
\concept{str-utils}
