% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str_starts}
\alias{str_starts}
\alias{str_ends}
\title{str_starts}
\usage{
str_starts(string, pattern, negate = FALSE)

str_ends(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{Input string}

\item{pattern}{Pattern to match}

\item{negate}{If \code{TRUE}, return non-matching elements}
}
\value{
a logical vector
}
\description{
Detect patterns at the beginning or end of strings
}
\seealso{
Other str-utils: 
\code{\link{str_detect}()},
\code{\link{str_extract}()},
\code{\link{str_wrap}()}
}
\author{
Guangchuang Yu
}
\concept{str-utils}
