% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utilities.R
\name{str_starts}
\alias{str_starts}
\alias{str_ends}
\title{str_starts}
\usage{
str_starts(string, pattern, negate = FALSE)

str_ends(string, pattern, negate = FALSE)
}
\arguments{
\item{string}{input string}

\item{pattern}{pattern with which the string starts or ends}

\item{negate}{if TRUE, return non-matching elements}
}
\value{
a logical vector
}
\description{
Detect the presence or absence of a pattern at the beginning or end of a string or string vector.
}
\author{
Guangchuang Yu
}
