% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-wrap.R
\name{str_wrap}
\alias{str_wrap}
\title{str_wrap}
\usage{
str_wrap(string, width = getOption("width"))
}
\arguments{
\item{string}{input string}

\item{width}{the maximum number of characters before wrapping to a new line}
}
\value{
update strings with new line character inserted
}
\description{
wraping long string to multiple lines
}
\author{
Guangchuang Yu and Erqiang Hu
}
