% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{user_dir}
\alias{user_dir}
\title{user_dir}
\usage{
user_dir(appname = NULL, appauthor = NULL, ...)
}
\arguments{
\item{appname}{App name}

\item{appauthor}{App author}

\item{...}{additional parameters}
}
\value{
a directory (created if not exists)
}
\description{
get the user dir to save app caches, logs and data (a wrapper function of \code{rappdirs::user_cache_dir()})
}
\author{
Guangchuang Yu
}
