% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{yread_tsv}
\alias{yread_tsv}
\alias{yread}
\title{yread}
\usage{
yread_tsv(
  file,
  reader = utils::read.delim,
  params = list(),
  cache_dir = tempdir()
)

yread(file, reader = readLines, params = list(), cache_dir = NULL)
}
\arguments{
\item{file}{a file or url}

\item{reader}{a function to read the 'file_url'}

\item{params}{a list of parameters that passed to the 'reader'}

\item{cache_dir}{a folder to store cache files. If set to NULL will disable cache.}
}
\value{
the output of using the 'reader' to read the 'file_url' with parameters specified by the 'params'
}
\description{
read file with caching
}
\details{
This function read a file (local or url) and cache the content.
}
\seealso{
Other io-utils: 
\code{\link{o}()},
\code{\link{read.cb}()},
\code{\link{show_in_excel}()}

Other io-utils: 
\code{\link{o}()},
\code{\link{read.cb}()},
\code{\link{show_in_excel}()}
}
\author{
Yonghe Xia and Guangchuang Yu
}
\concept{io-utils}
