% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yulab-msg.R
\name{yulab_msg}
\alias{yulab_msg}
\title{yulab_msg}
\usage{
yulab_msg(pkgname = NULL, n = 1)
}
\arguments{
\item{pkgname}{Package name}

\item{n}{Number of citation messages}
}
\value{
Package message
}
\description{
Messages for YuLab packages
}
\author{
Guangchuang Yu
}
